Move common GetRequestsFromBatch and GetTokenizedUrl methods to GraphUtils
