+++ DevProxy.Plugins/Utils/GraphUtils.cs
@@ -2,9 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using DevProxy.Abstractions.Models;
+using DevProxy.Abstractions.Utils;
 using DevProxy.Plugins.Models;
 using Microsoft.Extensions.Logging;
 using System.Net.Http.Json;
+using System.Text.Json;
 using Titanium.Web.Proxy.Http;
 
 namespace DevProxy.Plugins.Utils;
@@ -101,7 +104,7 @@ sealed class GraphUtils(
 
     public static MethodAndUrl GetMethodAndUrl(string methodAndUrlString)
     {
-        ArgumentException.ThrowIfNullOrWhiteSpace(methodAndUrlString, nameof(methodAndUrlString));
+        ArgumentException.ThrowIfNullOrWhiteSpace(methodAndUrlString);
 
         var info = methodAndUrlString.Split(" ");
         if (info.Length > 2)
@@ -111,4 +114,44 @@ sealed class GraphUtils(
         return new(Method: info[0], Url: info[1]);
     }
 
+    public static string GetTokenizedUrl(string absoluteUrl)
+    {
+        var sanitizedUrl = ProxyUtils.SanitizeUrl(absoluteUrl);
+        return "/" + string.Concat(new Uri(sanitizedUrl).Segments.Skip(2).Select(Uri.UnescapeDataString));
+    }
+
+    public static MethodAndUrl[] GetRequestsFromBatch(string batchBody, string graphVersion, string graphHostName)
+    {
+        var requests = new List<MethodAndUrl>();
+
+        if (string.IsNullOrEmpty(batchBody))
+        {
+            return [.. requests];
+        }
+
+        try
+        {
+            var batch = JsonSerializer.Deserialize<GraphBatchRequestPayload>(batchBody, ProxyUtils.JsonSerializerOptions);
+            if (batch == null)
+            {
+                return [.. requests];
+            }
+
+            foreach (var request in batch.Requests)
+            {
+                try
+                {
+                    var method = request.Method;
+                    var url = request.Url;
+                    var absoluteUrl = $"https://{graphHostName}/{graphVersion}{url}";
+                    MethodAndUrl methodAndUrl = new(Method: method, Url: GetTokenizedUrl(absoluteUrl));
+                    requests.Add(methodAndUrl);
+                }
+                catch { }
+            }
+        }
+        catch { }
+
+        return [.. requests];
+    }
 }
